feat: rename 'session' to 'space'
